requirements.txt
-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
base58==2.1.0; python_version >= '3.5'
bitarray==1.2.2
certifi==2021.10.8
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
cytoolz==0.11.0; implementation_name == 'cpython'
eth-abi==2.1.1; python_version >= '3.6' and python_version < '4'
eth-account==0.5.6; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.3.2; python_version >= '3.5' and python_version < '4'
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4'
eth-typing==2.2.2; python_version >= '3.5' and python_version < '4'
eth-utils==1.10.0; python_version >= '3.5' and python_version < '4' and python_full_version != '3.5.2'
hexbytes==0.2.2; python_version >= '3.6' and python_version < '4'
idna==3.3; python_version >= '3'
ipfshttpclient==0.8.0a2; python_full_version >= '3.6.2' and python_full_version not in '3.7.0, 3.7.1'
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==5.2.0; python_version >= '3.6'
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.1; python_version >= '3.5'
pycryptodome==3.11.0
pyrsistent==0.18.0; python_version >= '3.6'
python-dotenv==0.19.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rlp==2.0.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.11.1; python_version >= '3.5'
typing-extensions==3.10.0.2
uniswap-python==0.5.3
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
varint==1.0.2
web3==5.24.0; python_version >= '3.6' and python_version < '4'
websockets==9.1; python_full_version >= '3.6.1'
yarl==1.7.0; python_version >= '3.6'
